Smartlabs Agustín y Uriel IoT/requirements.txt
Flask
matplotlib
numpy
scipy
pyserial
